requirements.txt
streamlit
firebase-admin
google-genai
yagmail